requirements.txt
beautifulsoup4==4.8.2
blis==0.4.1
boto==2.49.0
boto3==1.10.42
botocore==1.13.42
bs4==0.0.1
catalogue==0.0.8
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
dj-database-url==0.5.0
Django==2.2.5
django-appconf==1.0.3
django-bootstrap-modal-forms==1.4.4
django-crispy-forms==1.7.2
django-sb-admin==0.3.7
django-widget-tweaks==1.4.5
docutils==0.15.2
fuzzywuzzy==0.17.0
gensim==3.8.1
google==2.0.3
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
importlib-metadata==1.3.0
jmespath==0.9.4
libsass==0.19.3
more-itertools==8.0.2
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.4
Pillow==6.1.0
plac==1.1.3
preshed==3.0.2
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.3
reportlab==3.5.34
requests==2.22.0
rjsmin==1.1.0
s3transfer==0.2.1
scipy==1.4.1
six==1.12.0
smart-open==1.9.0
soupsieve==1.9.5
sqlparse==0.3.0
srsly==0.2.0
thinc==7.3.1
tqdm==4.40.2
urllib3==1.25.7
wasabi==0.4.2
webencodings==0.5.1
whitenoise==4.1.4
wincertstore==0.2
zipp==0.6.0
